requirements.txt
pandas
numpy 
matplotlib.pyplot 
seaborn 
re
os
sklearn
tqdm
nltk
warnings
cv2
sklearn.model_selection
train_test_split
PIL
PIL 
time
tensorflow
keras
keras.layers
tensorflow.keras
tensorflow.keras.layers
nltk.translate.bleu_score 
tensorflow.keras.models
pickle